fm-alarm-lib/requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
contextlib2==0.6.0.post1
coverage==4.5.4
enum34==1.1.6
funcsigs==1.0.2
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
magicmock==0.3
mock==3.0.5
more-itertools==5.0.0
packaging==19.2
pathlib2==2.3.5
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.4
pytest==4.6.6
pytest-cov==2.8.1
python-consul==1.1.0
requests==2.22.0
requests-mock==1.7.0
scandir==1.10.0
six==1.13.0
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
